finalize features + security auth

+++ back/src/main/java/com/openclassrooms/mddapi/repository/ArticleRepository.java
@@ -6,4 +6,4 @@ import org.springframework.stereotype.Repository;
 import com.openclassrooms.mddapi.models.Article;
 
 @Repository
-public interface ArticleRepository extends JpaRepository<Article, Long> { }
+public interface ArticleRepository extends JpaRepository<Article, Integer> { }
